# frontend/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "RouteRank API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/admin/scoring/run" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "runScoring",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reason"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Run"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/routes/{id}/restore" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "restoreRoute",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reason"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/admin/routes/{id}/remove" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "removeRoute",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reason"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/admin/clusters/remove" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "removeCluster",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClusterAction"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Resolved"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/clusters/release" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "releaseCluster",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClusterAction"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Resolved"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/accounts/{id}/unban" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "unban",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reason"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/admin/accounts/{id}/release" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "releaseAccount",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reason"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/admin/accounts/{id}/ban" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "ban",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reason"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/admin/audit" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "auditLog",
        "parameters" : [ {
          "name" : "before",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Entry"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/accounts" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "findAccounts",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AccountRow"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/accounts/{id}" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "account",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Reason" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          }
        },
        "required" : [ "reason" ]
      },
      "Run" : {
        "type" : "object",
        "properties" : {
          "at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "segments" : {
            "type" : "integer",
            "format" : "int32"
          },
          "stretches" : {
            "type" : "integer",
            "format" : "int32"
          },
          "ranked" : {
            "type" : "integer",
            "format" : "int32"
          },
          "uploaded" : {
            "type" : "integer",
            "format" : "int32"
          },
          "deleted" : {
            "type" : "integer",
            "format" : "int32"
          },
          "published" : {
            "type" : "boolean"
          },
          "note" : {
            "type" : "string"
          }
        }
      },
      "ClusterAction" : {
        "type" : "object",
        "properties" : {
          "cluster" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "reason" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          }
        },
        "required" : [ "cluster", "reason" ]
      },
      "Resolved" : {
        "type" : "object",
        "properties" : {
          "users" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "removedRoutes" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      },
...
      "RouteSummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "slot" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "busNumber" : {
            "type" : "string"
          },
          "lengthM" : {
            "type" : "number",
            "format" : "double"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "removedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "Entry" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "actorId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "actorEmail" : {
            "type" : "string"
          },
          "action" : {
            "type" : "string"
          },
          "target" : {
            "type" : "string"
          },
          "before" : {
            "type" : "string"
          },
          "after" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "at" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "AccountRow" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "liveAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "heldAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "bannedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "trustScore" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
      "AccountDetail" : {
        "type" : "object",
        "properties" : {
          "account" : {
            "$ref" : "#/components/schemas/AccountRow"
          },
          "routes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RouteSummary"
            }
          }
        }
      }
    }
  }
}